refactor package name from automaticLotto to lotto
