+++ src/main/java/lotto/domain/LottoGenerator.java
@@ -1,4 +1,4 @@
-package automaticLotto.domain;
+package lotto.domain;
 
 import java.util.ArrayList;
 import java.util.Collections;
