+++ src/main/java/lotto/domain/Lottos.java
@@ -1,4 +1,4 @@
-package automaticLotto.domain;
+package lotto.domain;
 
 import java.util.List;
 import java.util.function.Consumer;
