+++ src/main/java/lotto/domain/ProfitRate.java
@@ -1,4 +1,4 @@
-package automaticLotto.domain;
+package lotto.domain;
 
 public class ProfitRate {
 	double profitRate;
