+++ src/main/java/lotto/domain/RankingCondition.java
@@ -1,4 +1,4 @@
-package automaticLotto.domain;
+package lotto.domain;
 
 import java.util.Objects;
