+++ src/main/java/lotto/domain/WinnerTable.java
@@ -1,4 +1,4 @@
-package automaticLotto.domain;
+package lotto.domain;
 
 import java.util.HashMap;
 import java.util.Map;
