+++ src/main/java/lotto/view/InputView.java
@@ -1,4 +1,4 @@
-package automaticLotto.view;
+package lotto.view;
 
 import java.util.Arrays;
 import java.util.List;
