+++ src/test/java/lotto/domain/WinnerStaticsTest.java
@@ -1,4 +1,4 @@
-package automaticLotto.domain;
+package lotto.domain;
 
 import static org.assertj.core.api.Assertions.*;
